[SPMD] Split the SPMD flow to tracing, profiling and optimization stages
ghstack-source-id: 4093e96740584889f1a016e45e0c55c39f55737b
Pull Request resolved: https://github.com/pytorch/tau/pull/652
